BCEL-343 JUnit Assertion improvement (#69)
* BCEL-343 Fix assertion typo

Fix a typo in FieldAnnotationsTestCase#checkValue: "Didnt" to
"Didn't".

* BCEL-343 Remove unused test methods

* BCEL-343 Remove redundant casts in tests

* BCEL-343 Simplify assertTrue to assertFalse

Simplify assertTrue calls with negative conditions to assertFalse
calls which are easier to understand.

* BCEL-343 Simplify assertTrue with equality to assertEquals

Clean up assertions code by using the built-in assertions in
org.junit.jupiter.api.Assertions instead of reinventing the wheel.

assertTrue(x.equals(y)) and assertTrue(primitiveX == primitiveY)
are simplified to assertEquals(x, y).

The main gain by this patch is cleaning up the code and making it
easier to maintain.
A side bonus is that many assertions simplified to assertEquals calls
had messages that were manually constructed to show the difference
between the two values being compared, and these messages were
constructed regardless of the assertion success or failure. By using
assertEquals, these string concatenations are avoided as long as the
assert succeeds.

* BCEL-343 Simplify explicit ifs to assertEquals

Clean up assertions code by using the built-in assertions in
org.junit.jupiter.api.Assertions instead of reinventing the wheel.

This patch simplifies the idiom of using an if to explicitly check for
equality:
```
if (!a.equals(b)) {
    fail("a should equal b");
}
```

with the build it assertEquals(a, b, "a should equal b").

* BCEL-343 Simplify catching exceptions in tests

Replace the idiom of catching an exception and then explicitly calling
org.junit.jupiter.api.Assertions with just allowing the error to be
thrown from the test method and leaving the it up to the runner to
handle the failure.

* BCEL-343 Replace non-constant assertion message with Supplier

In the cases that assertions have non-constant failure messages, these
messages are replaced with Suppliers that produce them in order to
avoid producing the calculated string unconditionally.
Former-commit-id: 215508b60f372891d2da0c7190451d8e82bb404b